requirements.txt
fastapi
uvicorn[standard]
Pillow
transparent-background
python-multipart
python-dotenv